Remove sql statements
